refactor(permissions): centralize into config module
Moved all custom permission classes from individual apps
into a single permissions.py module under the lazydog_api
project config directory.

This was done to address Pylint R0801 (duplicate code)
violations raised in CI, and to better follow the DRY
principle across the codebase.
