--- comment/permissions.py
@@ -1,14 +1,0 @@
-from rest_framework import permissions
-
-
-class IsOwnerOrReadOnly(permissions.BasePermission):
-    """
-    Custom permission to only allow owners of a resource to edit or delete it.
-    """
-
-    def has_object_permission(self, request, view, obj):
-
-        if request.method in permissions.SAFE_METHODS:
-            return True
-
-        return obj.user == request.user
