+++ flag/views.py
@@ -1,27 +1,24 @@
 # flag/views.py
-from rest_framework import viewsets, permissions, filters
+from rest_framework import viewsets, filters
 from django_filters.rest_framework import DjangoFilterBackend
 from .models import Flag
 from .serializers import FlagSerializer
-from .permissions import IsAdminOrOwner
+from lazydog_api.permissions import IsAdminOrOwner
 
 
 class FlagViewSet(viewsets.ModelViewSet):
     queryset = Flag.objects.all()
     serializer_class = FlagSerializer
-    permission_classes = [
-        permissions.IsAuthenticatedOrReadOnly,
-        IsAdminOrOwner,
-    ]
+    permission_classes = [IsAdminOrOwner]
     filter_backends = [
         DjangoFilterBackend,
         filters.OrderingFilter,
         filters.SearchFilter,
     ]
-    filterset_fields = ['user','resource', 'comment', 'status']
-    search_fields = ['reason']
-    ordering_fields = ['created_at', 'status']
-    ordering = ['-created_at']
+    filterset_fields = ["user", "resource", "comment", "status"]
+    search_fields = ["reason"]
+    ordering_fields = ["created_at", "status"]
+    ordering = ["-created_at"]
 
     def perform_create(self, serializer):
         serializer.save(user=self.request.user)
